[SCAL-200399] Remove redundant calls to TS create/update APIs during User, Group Sync (org_ldap_sync_python_api)
